Move wxWidgetWindowPlacement to separate files, publicize GetParent()
