+++ src/export/ExportPCM.cpp
@@ -34,7 +34,7 @@
 #include "../Track.h"
 #include "../widgets/AudacityMessageBox.h"
 #include "../widgets/ProgressDialog.h"
-#include "../widgets/wxWidgetsBasicUI.h"
+#include "../widgets/wxWidgetsWindowPlacement.h"
 #include "wxFileNameWrapper.h"
 
 #include "Export.h"
